requirements.txt
absl-py==2.2.0
array_record==0.7.1
astunparse==1.6.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
dm-tree==0.1.9
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
flatbuffers==25.2.10
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.21.0
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
promise==2.3
protobuf==3.20.3
psutil==7.0.0
pyarrow==19.0.1
Pygments==2.19.1
requests==2.32.3
rich==13.9.4
simple-parsing==0.1.7
six==1.17.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.8
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
termcolor==2.5.0
tf2onnx==1.16.1
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
